requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-versioning==2.1.0
drf-link-header-pagination==0.2.0
packaging==23.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing==3.7.4.3
typing_extensions==4.8.0
urllib3==2.0.7
